Retry most requests on s3 errors

+++ boto_addins/s3.py
@@ -29,16 +29,9 @@ class AsyncS3Connection(S3Connection):
         return super(AsyncS3Connection, self).get_bucket(
             bucket_name, validate, headers)
 
-    def fetch_request(self, request,
-                      client=simple_httpclient.SimpleAsyncHTTPClient):
-        """
-        For test purposes mainly.
-        """
-        return client().fetch(request)
-
     def generate_async_request(self, method, bucket='', key='',
                                query_args=None, headers=None,
-                               content_length=None, **kwargs):
+                               content_length=None, attempts=1, **kwargs):
         if isinstance(bucket, self.bucket_class):
             bucket = bucket.name
         if isinstance(key, Key):
@@ -63,11 +56,12 @@ class AsyncS3Connection(S3Connection):
             "{0.protocol}://{0.host}{0.path}".format(req),
             req.method, req.headers, **kwargs
         )
-        return self.fetch_request(async_request)
+        return fetch_request(async_request, attempts=attempts)
 
     def async_set_contents_from_file(self, bucket, key_name, fp, headers=None,
                                      policy=None, encrypt_key=False,
-                                     metadata=None):
+                                     metadata=None, attempts=3,
+                                     request_timeout=10 * 60):
         headers = headers or {}
         if policy:
             headers[self.provider.acl_header] = policy
@@ -78,11 +72,13 @@ class AsyncS3Connection(S3Connection):
 
         return self.generate_async_request(
             'PUT', bucket, key_name, headers=headers, body=fp.read(),
+            request_timeout=request_timeout, attempts=attempts,
         )
 
     @gen.coroutine
     def async_copy_key(self, src_key, dst_key, temp_dir, metadata=None,
-                       policy=None, encrypt_key=False, headers=None):
+                       policy=None, encrypt_key=False, headers=None,
+                       request_timeout=10 * 60):
         headers = headers or {}
         if policy:
             headers[self.provider.acl_header] = policy
@@ -93,7 +89,8 @@ class AsyncS3Connection(S3Connection):
 
         url = src_key.generate_url(S3_TEMP_URL_TTL)
         destination = os.path.join(temp_dir, '_' + str(uuid4()))
-        yield async_http_download(url, destination)
+        yield async_http_download(url, destination,
+                                  request_timeout=request_timeout)
 
         @gen.coroutine
         def producer(write):
@@ -109,7 +106,8 @@ class AsyncS3Connection(S3Connection):
                 'PUT', dst_key.bucket.name, dst_key.name,
                 headers=headers, body_producer=producer,
                 content_length=src_key.size,
-                request_timeout=10 * 60,  # Timeout for whole request, not tcp.
+                # Timeout for whole request, not tcp.
+                request_timeout=request_timeout,
             )
         finally:
             if os.path.exists(destination):
@@ -121,14 +119,14 @@ class AsyncBucket(Bucket):
         self._downloading_files = weakref.WeakValueDictionary()
         super(AsyncBucket, self).__init__(connection, name, key_class=AsyncKey)
 
-    def async_get_key_contents(self, key_name, **kwargs):
+    def async_get_key_contents(self, key_name, attempts=3, **kwargs):
         return self.connection.generate_async_request(
-            'GET', self.name, key_name, **kwargs
+            'GET', self.name, key_name, attempts=attempts, **kwargs
         )
 
     @gen.coroutine
     def async_get_key(self, key_name, headers=None, version_id=None,
-                      response_headers=None):
+                      response_headers=None, attempts=3):
         query_args_l = []
         if version_id:
             query_args_l.append('versionId=%s' % version_id)
@@ -140,7 +138,7 @@ class AsyncBucket(Bucket):
         try:
             response = yield self.connection.generate_async_request(
                 'HEAD', self.name, key_name,
-                headers=headers, query_args=query_args,
+                headers=headers, query_args=query_args, attempts=attempts,
             )
         except httpclient.HTTPError as e:
             if e.code == 404:
@@ -153,13 +151,15 @@ class AsyncBucket(Bucket):
         raise gen.Return(key)
 
     @gen.coroutine
-    def async_get_acl(self, key_name='', headers=None, version_id=None):
+    def async_get_acl(self, key_name='', headers=None, version_id=None,
+                      attempts=3):
         query_args = 'acl'
         if version_id:
             query_args += '&versionId=%s' % version_id
 
         resp = yield self.connection.generate_async_request(
-            'GET', self.name, key_name, query_args=query_args, headers=headers
+            'GET', self.name, key_name, query_args=query_args, headers=headers,
+            attempts=attempts,
         )
 
         policy = Policy(self)
@@ -168,13 +168,16 @@ class AsyncBucket(Bucket):
         raise gen.Return(policy)
 
     @gen.coroutine
-    def async_download_key(self, key, path):
+    def async_download_key(self, key, path,
+                           request_timeout=10 * 60, attempts=3):
         # Get shared downloading.
         future = self._downloading_files.get(key.name)
 
         if future is None:
             url = key.generate_url(S3_TEMP_URL_TTL)
-            future = async_http_download(url, path)
+            future = async_http_download(
+                url, path, request_timeout=request_timeout, attempts=attempts,
+            )
 
             # Share downloading.
             self._downloading_files[key.name] = future
@@ -189,19 +192,67 @@ class AsyncKey(Key):
 
 
 @gen.coroutine
-def async_http_download(source, destination):
+def fetch_request(request, client=None, retry_callback=None, attempts=1):
+    if client is None:
+        client = simple_httpclient.SimpleAsyncHTTPClient()
+
+    # Fail faster on connection if we can retry
+    request.connect_timeout = 5 if attempts > 1 else 20
+
+    while attempts:
+        wait_before_retry = 0
+        try:
+            resp = yield client.fetch(request)
+        except httpclient.HTTPError as e:
+            # retry on s3 errors
+            if e.code == 500:
+                wait_before_retry = 0.2
+            elif e.code in (503, 599):
+                wait_before_retry = 1
+            else:
+                raise
+        else:
+            raise gen.Return(resp)
+
+        attempts -= 1
+        if not attempts:
+            raise
+
+        if wait_before_retry:
+            yield gen.sleep(wait_before_retry)
+        if retry_callback:
+            yield retry_callback(request, attempts)
+
+
+@gen.coroutine
+def async_http_download(source, destination,
+                        request_timeout=10 * 60, attempts=3):
     tmp_name = '{}.{}'.format(destination, str(uuid4())[:8])
 
     try:
         with open(tmp_name, 'w') as iobuffer:
-            yield curl_httpclient.CurlAsyncHTTPClient().fetch(
-                source,
-                streaming_callback=iobuffer.write,  # Direct write to file.
-                request_timeout=10 * 60,  # Timeout for whole request, not tcp.
+
+            @gen.coroutine
+            def reset_iobuffer(*args):
+                iobuffer.seek(0)
+                iobuffer.truncate()
+
+            yield fetch_request(
+                httpclient.HTTPRequest(
+                    source,
+                    # Direct write to file.
+                    streaming_callback=iobuffer.write,
+                    # Timeout for whole request, not tcp.
+                    request_timeout=request_timeout,
+                ),
+                client=curl_httpclient.CurlAsyncHTTPClient(),
+                retry_callback=reset_iobuffer,
+                attempts=attempts,
             )
 
     except BaseException:
-        os.unlink(tmp_name)
+        if os.path.exists(tmp_name):
+            os.unlink(tmp_name)
         raise
 
     # concurrent downloading
